+++ soil.php
@@ -50,6 +50,8 @@ class Options {
   }
 }
 
+require_once __DIR__ . '/lib/utils.php';
+
 function load_modules() {
   global $_wp_theme_features;
   foreach (glob(__DIR__ . '/modules/*.php') as $file) {